refactor(usecase): rename directory
