+++ package.json
@@ -45,6 +45,7 @@
     "babel-preset-react": "^6.5.0",
     "babel-register": "^6.7.2",
     "chokidar-cli": "^1.2.0",
+    "gh-pages": "^0.11.0",
     "json-loader": "^0.5.4",
     "mkdirp": "^0.5.1",
     "mocha": "^2.4.5",
