+++ src/js/UseCaseController/ExportUseCaseController.js
@@ -1,7 +1,7 @@
 // LICENSE : MIT
 "use strict";
 import documentRepository from "../infra/DocumentRepository";
-import ShowExportDialogUseCase from "../usecase/ShowExportDialogUseCase";
+import ShowExportDialogUseCase from "../UseCase/ShowExportDialogUseCase";
 export default class ExportUseCaseController {
     static ShowExportDialogUseCase() {
         var useCase = new ShowExportDialogUseCase({documentRepository});
