--- src/js/usecase/_port/InputPort.js
@@ -1,6 +1,0 @@
-// LICENSE : MIT
-"use strict";
-// Input Data -> UseCase#call -> data -> Output Port
-export default class InputPort {
-    
-}
