--- src/js/usecase/_port/OutputPort.js
@@ -1,5 +1,0 @@
-// LICENSE : MIT
-"use strict";
-export default class OutputPort {
-
-}
