--- src/js/usecase/_port/UseCaseInferface.js
@@ -1,10 +1,0 @@
-// LICENSE : MIT
-"use strict";
-export default class UseCaseInterface {
-    // call -+ return promise
-    // define input
-    // define output
-    call(...args) {
-        return Promise.resolve();
-    }
-}
