+++ test/usecase/CompleteLoadingDocumentUseCase-test.js
@@ -1,7 +1,7 @@
 // LICENSE : MIT
 "use strict";
 const assert = require("power-assert");
-import CompleteLoadingDocumentUseCase from "../../src/js/usecase/CompleteLoadingDocumentUseCase";
+import CompleteLoadingDocumentUseCase from "../../src/js/UseCase/CompleteLoadingDocumentUseCase";
 import Document from "../../src/js/domain/Document/Document";
 import {DocumentRepository} from "../../src/js/infra/DocumentRepository";
 describe("CompleteLoadingDocumentUseCase", function () {
